% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgLadder.R
\name{avgLadder}
\alias{avgLadder}
\title{Average ladder size}
\usage{
avgLadder(tree)
}
\arguments{
\item{tree}{a tree of class \code{phylo} or \code{phylo4}}
}
\value{
The mean ladder size
}
\description{
Finds the mean size of ladders in the tree
}
\examples{
## Find the average ladder size in a random tree with 20 tips:
avgLadder(rtree(20))


}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

