% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseAnalysis.R
\name{baseAnalysis}
\alias{baseAnalysis}
\title{Analyse base properties of trees}
\usage{
baseAnalysis(base, topFuncs)
}
\arguments{
\item{base}{an output from baseCreate}

\item{topFuncs}{a list of top functions}
}
\value{
A matrix of values
}
\description{
Called by \code{baseListAnalysis} to apply topFuncs to baseFunc outputs
}
\author{
Michael Boyd \email{mboyd855@gmail.com}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

