% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseCreate.R
\name{baseCreate}
\alias{baseCreate}
\title{List of basic tree functions}
\usage{
baseCreate(tree, baseFuncs)
}
\arguments{
\item{tree}{a tree of class \code{phylo4}}

\item{baseFuncs}{a list of base functions}
}
\value{
A list of base functions evaluated on this tree.
}
\description{
Creates a base list for the tree to be used later.
Notes: needs to issue warning about input types.
Adapt for phylo as well as phylo4?
}
\examples{
## Find the average ladder size and number of cherries in a random tree
baseCreate(rtree4(10),c(avgLadder,cherries))


}
\author{
Michael Boyd \email{mboyd855@gmail.com}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

