% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseListAnalysis.R
\name{baseListAnalysis}
\alias{baseListAnalysis}
\title{Analysis of a list of basic tree functions for a list of trees}
\usage{
baseListAnalysis(baseList, topFuncs, loadingBar = FALSE)
}
\arguments{
\item{baseList}{an object of the type created by baseListCreate.}

\item{topFuncs}{a list of functions that operate on elements of baseList.}

\item{loadingBar}{option to include a loading bar. Defaults to \code{FALSE}.}
}
\value{
Analysis of a list of base functions evaluated on each tree.
}
\description{
Takes a base list for a list of trees, and analyses it. 
Notes: needs to issue warning about input types.
Adapt for phylo as well as phylo4?
Name columns of output so you can see which topFuncs were used
}
\examples{
## Find the average ladder size and number of cherries in a random tree
testlist <- baseListCreate(c(rtree4(10),rtree4(12)),c(avgLadder,cherries))
baseListAnalysis(testlist,c(mean,max))

}
\author{
Michael Boyd \email{mboyd855@gmail.com}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

