% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseListCreate.R
\name{baseListCreate}
\alias{baseListCreate}
\title{List of basic tree functions for a list of trees}
\usage{
baseListCreate(treeList, baseFuncs, loadingBar = TRUE)
}
\arguments{
\item{treeList}{a list of trees of class \code{phylo4}}

\item{baseFuncs}{a list of base functions}
}
\value{
A list of base functions evaluated on each tree.
}
\description{
Creates a base list for each tree in a list.
}
\examples{
## Find the average ladder size and number of cherries in a random tree
baseListCreate(c(rtree4(10),rtree4(12)),c(avgLadder,cherries))


}
\author{
Michael Boyd \email{mboyd855@gmail.com}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

