% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cherries.R
\name{cherries}
\alias{cherries}
\title{Cherry number}
\usage{
cherries(tree)
}
\arguments{
\item{tree}{a tree of class \code{phylo4}}
}
\value{
An integer representing the number of cherries in the tree.
}
\description{
Finds the number of cherries in a tree
Notes: needs to issue warning about input types.
Adapt for phylo as well as phylo4?
}
\examples{
cherries(rtree4(10))


}
\author{
Michael Boyd \email{mboyd855@gmail.com}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

