% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colless.R
\name{colless}
\alias{colless}
\title{Colless number}
\usage{
colless(tree, normalize = TRUE)
}
\arguments{
\item{tree}{a tree of class \code{phylo4}}
}
\value{
The Colless imbalance number of the tree.
}
\description{
Finds the Colless number for a tree
Notes: needs to issue warning about input types.
Adapt for phylo as well as phylo4?
}
\examples{
colless(rtree4(10))


}
\author{
Michael Boyd \email{mboyd855@gmail.com}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

