% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configFind.R
\name{configFind}
\alias{configFind}
\title{Find configurations of a given size}
\usage{
configFind(tree, configSize)
}
\arguments{
\item{tree}{a tree of class \code{phylo4}}
}
\value{
An integer representing the number of cherries in the tree.
}
\description{
I THINK: find the internal node IDs whose number of tip descendants is equal to configSize. 
However, it didn't do this at first; I have added "type="internal"" to make it work as I expected...
Come back to this.
Notes: needs to issue warning about input types.
Adapt for phylo as well as phylo4?
}
\examples{
configFind(rtree4(10),2)


}
\author{
Michael Boyd \email{mboyd855@gmail.com}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

