% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configShow.R
\name{configShow}
\alias{configShow}
\title{Plot a tree highlighting configurations}
\usage{
configShow(tree, configSize)
}
\arguments{
\item{tree}{a tree of class \code{phylo4}}

\item{configSize}{an integer giving the configuration size of interest.}
}
\value{
An integer representing the number of cherries in the tree.
}
\description{
Plots the tree with configurations of a given size highlighted.
CURRENTLY BROKEN.
Probably needs argument "type="internal"" in nodeId?
Notes: needs to issue warning about input types.
Adapt for phylo as well as phylo4?
}
\examples{
## Highlight cherries in a random tree:
configShow(rtree4(10),2)


}
\author{
Michael Boyd \email{mboyd855@gmail.com}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

