% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configTop.R
\name{configTop}
\alias{configTop}
\title{Number of each type of configuration}
\usage{
configTop(tree)
}
\arguments{
\item{tree}{a tree of class \code{phylo4}}
}
\value{
A vector of the same length as the number of tips. Each entry corresponds to the number of configurations of that size.
For example, the second entry gives the number of configurations of size two; the nth entry gives the number of configurations of size n.
}
\description{
Finds the number of each type of configuration
Notes: needs to issue warning about input types.
Adapt for phylo as well as phylo4?
}
\examples{
configTop(rtree4(10))


}
\author{
Michael Boyd \email{mboyd855@gmail.com}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

