% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{dists}
\alias{dists}
\title{Gives the depth of each node}
\usage{
dists(tree)
}
\arguments{
\item{tree}{a tree of class \code{phylo4}}
}
\value{
A vector giving the depth of each node
}
\description{
DEPENDS ON A DEPRICATED FUNCTION. UPDATE.
Notes: needs to issue warning about input types.
Adapt for phylo as well as phylo4?
}
\examples{
dists(mytree)


}
\author{
Michael Boyd \email{mboyd855@gmail.com}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

