% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fAvgLadder.R
\name{fAvgLadder}
\alias{fAvgLadder}
\title{Average ladder size}
\usage{
fAvgLadder(ladderNumbers)
}
\arguments{
\item{ladderNumbers}{a vector of ladder numbers as output by \code{ladderNums}}
}
\value{
The average ladder size
}
\description{
Finds the mean size of ladders in a tree; unlike \code{avgLadder} this function
does not take a tree as input, but takes the ladder numbers themselves, as output by \code{ladderNums}
}
\examples{
## Find the average ladder size in a random tree
fAvgLadder(ladderNums(rtree4(10)))


}
\author{
Michael Boyd \email{mboyd855@gmail.com}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

