% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fColless.R
\name{fColless}
\alias{fColless}
\title{Colless number}
\usage{
fColless(treeImbalance)
}
\arguments{
\item{treeImbalance}{a tree imbalance matrix as output by \code{treeImb}}
}
\value{
An integer giving the Colless number the tree.
}
\description{
Finds the Colless number, given a tree imbalance as input
Notes: needs to issue warning about input types.
Adapt for phylo as well as phylo4?
}
\examples{
fColless(treeImb(mytree))


}
\author{
Michael Boyd \email{mboyd855@gmail.com}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

