% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idNodeLabel.R
\name{idNodeLabel}
\alias{idNodeLabel}
\title{Change labels of tree to integers}
\usage{
idNodeLabel(tree)
}
\arguments{
\item{tree}{a tree of class \code{phylo4}}
}
\value{
A tree of class \code{phylo4} with relabelled nodes.
}
\description{
A tree created by \code{rtree4(k)} will have tip labels "t1", "t2", ... , "tk" and the internal node labels will all be "<NA>".
This function will relabel the tips to 1, 2, ..., k, and the internal nodes to k+1, k+2, ... , 2k - 1.
NOTE that the tip numbering is lost, is this deliberate?
Notes: needs to issue warning about input types.
Adapt for phylo as well as phylo4?
}
\examples{
idNodeLabel(rtree(10))


}
\author{
Michael Boyd \email{mboyd855@gmail.com}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

