% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladderNums.R
\name{ladderNums}
\alias{ladderNums}
\title{Ladder numbers}
\usage{
ladderNums(tree)
}
\arguments{
\item{tree}{a tree of class \code{phylo} or \code{phylo4}}
}
\value{
A vector corresponding to tree edges and their ladder sizes
}
\description{
Finds the ladders and their positions in the tree, for plotting
}
\author{
Michael Boyd \email{mboyd855@gmail.com}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}
\keyword{internal}

