% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladderShow.R
\name{ladderShow}
\alias{ladderShow}
\title{Show ladders}
\usage{
ladderShow(tree)
}
\arguments{
\item{tree}{a tree of class \code{phylo} or \code{phylo4}}
}
\value{
A plot of the tree, with ladders highlighted in red.
}
\description{
Plot a tree, highlighting any ladders within it
}
\examples{
\dontrun{
ladderShow(rtree(100))
}

}
\author{
Michael Boyd \email{mboyd855@gmail.com}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

