% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nTipDescendants.R
\name{nTipDescendants}
\alias{nTipDescendants}
\title{Number of tip descendants}
\usage{
nTipDescendants(tree)
}
\arguments{
\item{tree}{a tree of class \code{phylo4}}
}
\value{
A vector whose entries correspond to internal nodes of the tree, and give the number of tip descendants of each node.
}
\description{
Finds the number of tip descendants for each internal node of the tree. Used by configFind and configShow.
}
\examples{
nTipDescendants(rtree4(10))

}
\author{
Michael Boyd \email{mboyd855@gmail.com}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}
\keyword{internal}

