% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloCheck.R
\name{phyloCheck}
\alias{phyloCheck}
\title{Check tree format (internal)}
\usage{
phyloCheck(tree)
}
\arguments{
\item{tree}{a tree of class \code{phylo} or \code{phylo4}. For most functions in this package the tree should be binary and rooted, hence this function is called to check.
If necessary the tree will be coerced into a binary rooted tree using multi2di, if possible.}
}
\value{
A binary, rooted tree of class \code{phylo}, if possible.
}
\description{
Coerces the tree into a rooted, binary tree of class \code{phylo}. Note that this function used to require trees to be of class \code{phylo4} but the package now uses class \code{phylo} throughout.
}
\author{
Caroline Colijn \email{c.colijn@imperial.ac.uk}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}
\keyword{internal}
