\name{phyloTop-package}
\alias{phyloTop-package}
\alias{phyloTop}
\docType{package}
\title{
Phylogenetic Tree Topological Properties Evaluator
}
\description{
Calculates some topological properties of phylogenetic trees.
}
\details{
\tabular{ll}{
Package: \tab phyloTop\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2013-10-05\cr
License: \tab GPL-2\cr
}
This package is based on the package \code{phylobase}. See its documentation for many useful related functions. In particular \code{nodeId} and \code{treewalk}. \code{library(help=phyloTop)} will give more information on the package. 

Many thanks to Caroline Colijn of Imperial College who provided some of the data accompanying the package and supervised me in the writing of it.
}

\note{
I recommend running \code{phylobase.options(allow.duplicated.labels='ok')} to remove some warning messages.
}
\author{
Michael Boyd

Maintainer: Michael Boyd <mboyd855@gmail.com>
}
