% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchforks.R
\name{pitchforks}
\alias{pitchforks}
\title{Number of pitchforks}
\usage{
pitchforks(tree)
}
\arguments{
\item{tree}{a tree of class \code{phylo} or \code{phylo4}}
}
\value{
An integer representing the number of pitchforks in the tree.
}
\description{
Finds the number of pitchforks in a tree. A pitchfork is considered to be a clade with three tips.
}
\examples{
tree <- rtree(10)
plot(tree)
pitchforks(tree)


}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

