% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtree4.R
\name{rtree4}
\alias{rtree4}
\title{Random phylo4 tree}
\usage{
rtree4(n, ...)
}
\arguments{
\item{n}{an integer giving the number of tips in the tree}

\item{...}{further arguments to be passed to rtree}
}
\value{
An object of class \code{"phylo4"}.
}
\description{
Creates a random phylo4 tree, in parallel to \code{ape}'s \code{rtree} function.
}
\examples{
## Create a random phylo4 tree with 10 tips:
tree4 <- rtree4(10)


}
\author{
Michael Boyd \email{mboyd855@gmail.com}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}
