\name{rTrees}
\alias{rTrees}
\docType{data}
\title{
Random Trees
}
\description{
A list of 50 trees generated by \code{rtree4}.
}
\usage{data(rTrees)}
\format{
A list of length 50 where each object is an object of class \code{phylo4}.
}
\source{
Created by Michael Boyd.
}
\seealso{
\code{\link{exampleSummary}} for a more detailed summary.
}
\examples{
data(rTrees)
\dontrun{
topList <- c(avgLadder,colless,function(tree){nConfig(tree,2)})
treeListSummary(rTrees,topList)
}
}
\keyword{datasets}
