% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitTop.R
\name{splitTop}
\alias{splitTop}
\title{Split topology}
\usage{
splitTop(tree, dist)
}
\arguments{
\item{tree}{a tree of class \code{phylo} or \code{phylo4}. The tree should be binary and rooted; if not it will be coerced into a binary rooted tree using multi2di, if possible.}

\item{dist}{integer distance of nodes of interest from the root}
}
\value{
The mean ladder size
}
\description{
For nodes at a given distance from the root, this function finds their numbers of tip descendants
}
\examples{
## Find the split topology of a random tree with 20 tips:
tree <- rtree(20)
plot(tree)
splitTop(tree,2)


}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}

Michael Boyd \email{mboyd855@gmail.com}
}

