% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtreeShow.R
\name{subtreeShow}
\alias{subtreeShow}
\title{Highlight a subtree}
\usage{
subtreeShow(tree, nodeList)
}
\arguments{
\item{tree}{a tree of class \code{phylo} or \code{phylo4}.}

\item{nodeList}{a list of nodes in the tree}
}
\value{
A plot of the tree, with the specified clade(s) highlighted.
}
\description{
Plot a tree, highlighting the clade(s) descending from a specified node or nodes
}
\examples{
## Highlight the clade descending from node 13 in a random tree on 10 tips:
\dontrun{
subtreeShow(rtree(10),13)
}

}
\author{
Michael Boyd \email{mboyd855@gmail.com}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

