% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cherries.R
\name{cherries}
\alias{cherries}
\title{Cherry number}
\usage{
cherries(tree)
}
\arguments{
\item{tree}{a tree of class \code{phylo} or \code{phylo4}}
}
\value{
An integer representing the number of cherries in the tree.
}
\description{
Finds the number of cherries in a tree. A cherry is considered to be a pair of sister tips.
}
\examples{
tree <- rtree(10)
plot(tree)
cherries(tree)


}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

