% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configFind.R
\name{configFind}
\alias{configFind}
\title{Find configurations of a given size}
\usage{
configFind(tree, configSize)
}
\arguments{
\item{tree}{a tree of class \code{phylo} or \code{phylo4}}

\item{configSize}{integer giving size of configurations of interest}
}
\value{
The internal node IDS with the given number of tip descendants
}
\description{
Find the internal node IDs whose number of tip descendants is equal to configSize.
}
\author{
Michael Boyd \email{mboyd855@gmail.com}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}
\keyword{internal}

