% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILnumber.R
\name{ILnumber}
\alias{ILnumber}
\title{IL number}
\usage{
ILnumber(tree)
}
\arguments{
\item{tree}{a tree of class \code{phylo4}}
}
\value{
The integer number of internal nodes with a single tip child.
}
\description{
Computes the number of internal nodes with a single tip child. 
Notes: needs to issue warning about input types.
Adapt for phylo as well as phylo4?
}
\examples{
#ILnumber(rtree4(10))


}
\author{
Michael Boyd \email{mboyd855@gmail.com}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

