% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladderNums.R
\name{ladderNums}
\alias{ladderNums}
\title{Ladder numbers}
\usage{
ladderNums(tree)
}
\arguments{
\item{tree}{a tree of class \code{phylo4}}
}
\value{
A vector of ladder numbers (clarify)
}
\description{
Finds the ladder numbers
Notes: needs to issue warning about input types.
Adapt for phylo as well as phylo4?
}
\examples{
ladderNums(rtree4(10))


}
\author{
Michael Boyd \email{mboyd855@gmail.com}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

