% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladderShow.R
\name{ladderShow}
\alias{ladderShow}
\title{Show ladders}
\usage{
ladderShow(tree, mainCol = "black", ladderCol = "red", ...)
}
\arguments{
\item{tree}{a tree of class \code{phylo} or \code{phylo4}}

\item{mainCol}{colour for branches which are not ladders (default is black)}

\item{ladderCol}{colour for ladder branches (default is red)}

\item{...}{further arguments to be passed to plot.phylo}
}
\value{
A plot of the tree, with ladders highlighted by colour.
}
\description{
Plot a tree, highlighting any ladders within it
}
\examples{
## Highlight in blue the ladders in a random tree with 50 tips:
\dontrun{
tree <- rtree(50)
ladderShow(tree, ladderCol="blue", edge.width=2)
# compare to:
ladderSizes(tree)
}

}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}

Michael Boyd \email{mboyd855@gmail.com}
}
\seealso{
\code{\link{ladderSizes}}
}

