% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxHeight.R
\name{maxHeight}
\alias{maxHeight}
\title{Maximum tree height}
\usage{
maxHeight(tree, normalise = FALSE)
}
\arguments{
\item{tree}{a tree of class \code{phylo4}}

\item{normalise}{option to normalise the result, default is \code{FALSE}}
}
\value{
An integer giving the maximum height of tips in the tree.
}
\description{
Find the maximum height of tips in the tree.
}
\examples{
## Maximum height of tips in a random tree:
tree <- rtree(10)
maxHeight(tree) 
maxHeight(tree, normalise=TRUE)

}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}

Caroline Colijn \email{c.colijn@imperial.ac.uk}
}

