% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxHeight.R
\name{maxHeight}
\alias{maxHeight}
\title{Maximum tree height}
\usage{
maxHeight(tree)
}
\arguments{
\item{tree}{a tree of class \code{phylo4}}
}
\value{
An integer giving the maximum height of tips in the tree.
}
\description{
Find the maximum height of tips in the tree.
}
\examples{
## Maximum height of tips in a random tree:
maxHeight(rtree(10))


}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}

Caroline Colijn \email{c.colijn@imperial.ac.uk}
}

