% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nConfig.R
\name{nConfig}
\alias{nConfig}
\title{Number of internal nodes with a given number of tip descendants}
\usage{
nConfig(tree, configSize)
}
\arguments{
\item{tree}{a tree of class \code{phylo4}}

\item{configSize}{an integer giving the size of configuration of interest.}
}
\value{
An integer representing the number of internal nodes whose number of tip descendants is equal to \code{configSize}.
}
\description{
Finds the number of internal nodes whose number of tip descendants is equal to \code{configSize}.
Notes: needs to issue warning about input types.
Adapt for phylo as well as phylo4?
}
\examples{
## Find the number of cherries:
nConfig(rtree4(10),2)
## Find the number of pitchforks:
nConfig(rtree4(10),3)


}
\author{
Michael Boyd \email{mboyd855@gmail.com}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

