% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloTop.R
\name{phyloTop}
\alias{phyloTop}
\title{phyloTop: calculating topological properties of phylogenies}
\usage{
phyloTop(treeList, funcs = NULL)
}
\arguments{
\item{treeList}{a \code{list} or \code{multiPhylo} object, or a single tree of class \code{phylo} or \code{phylo4}. All trees should be binary and rooted; if not they will be coerced into binary rooted trees using multi2di, if possible.}

\item{funcs}{a list of functions. The default is to apply all of the topological functions from the package, but a subset can be specified instead. The functions available are:
\itemize{
\item avgLadder
\item cherries
\item colless.phylo
\item ILnumber
\item maxHeight
\item pitchforks
\item sackin.phylo
\item stairs (note that this adds two columns to the output, "stairs1" and "stairs2")
}}
}
\value{
A matrix where rows correspond to trees and columns correspond to topological properties.
}
\description{
Calculate a range of topological properties for one of more phylogenetic trees.
}
\examples{
## Apply all of the functions to a list of 10 random trees, each with 10 tips:
phyloTop(rmtree(10,10))


}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

