% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootLaddDist.R
\name{rootLaddDist}
\alias{rootLaddDist}
\title{Root ladder distance}
\usage{
rootLaddDist(tree)
}
\arguments{
\item{tree}{a tree of class \code{phylo4}}
}
\value{
Root ladder distance
}
\description{
Called by ladderNums
}
\author{
Michael Boyd \email{mboyd855@gmail.com}
}
\keyword{internal}

