% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stairs.R
\name{stairs}
\alias{stairs}
\title{Stairs}
\usage{
stairs(tree)
}
\arguments{
\item{tree}{a tree of class \code{phylo} or \code{phylo4}. The tree should be binary and rooted; if not it will be coerced into a binary rooted tree using multi2di, if possible.}
}
\value{
The stair numbers for a tree
}
\description{
Calculates the stair numbers
}
\examples{
## Find the stair numbers in a random tree with 20 tips:
stairs(rtree(20))
 

}
\author{
Michael Boyd \email{mboyd855@gmail.com}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

