% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeImb.R
\name{treeImb}
\alias{treeImb}
\title{Tree imbalance}
\usage{
treeImb(tree)
}
\arguments{
\item{tree}{a tree of class \code{phylo4}}
}
\value{
A matrix where columns correspond to nodes and rows correspond to descendant branches
}
\description{
Finds the imbalance of descendants: for each node it gives the number of descendants from of each descendant branch.
Notes: needs to issue warning about input types.
Does it work for non-binary trees?
Does it need to output a load of zeroes for tips or can we drop these entries?
Adapt for phylo as well as phylo4?
}
\examples{
treeImb(rtree4(10))


}
\author{
Michael Boyd \email{mboyd855@gmail.com}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

